requirements.txt
google-generativeai
pydantic
openai
langgraph
langchain
langchain-chroma
langchain-google-genai
chromadb
pypdf
